# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Board API",
    "version": "1.0.0"
  },
  "paths": {
    "/admin/jobs": {
      "post": {
        "tags": [
          "Jobs - Admin"
        ],
        "summary": "STEP 2 - Create a new job",
        "description": "Create job metadata. Job is created as INACTIVE by default and not visible to users.\nYou must add fields and then publish it to make it visible.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              },
              "example": {
                "title": "Frontend Developer",
                "description": "We are hiring a React developer with 3+ years of experience",
                "location": "Addis Ababa, Ethiopia",
                "type": "full-time",
                "work_mode": "hybrid",
                "key_responsibilities": [
                  "Develop and maintain web applications",
                  "Collaborate with cross-functional teams",
                  "Write clean, scalable code",
                  "Participate in code reviews",
                  "Troubleshoot and debug issues"
                ],
                "what_we_offer": [
                  "Competitive salary",
                  "Health and dental insurance",
                  "Remote work flexibility",
                  "Professional development opportunities",
                  "Generous vacation policy"
                ],
                "requirements": [
                  "Bachelor's degree in Computer Science or related field",
                  "3+ years of experience in software development",
                  "Proficiency in JavaScript, React, and Node.js",
                  "Strong problem-solving skills",
                  "Excellent communication and teamwork abilities"
                ],
                "deadline": "2026-02-15T17:00:00Z",
                "validFrom": "2026-01-12T10:00:00Z",
                "validTo": "2026-01-31T23:59:59Z"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                },
                "example": {
                  "success": true,
                  "message": "Job created successfully",
                  "data": {
                    "id": "677a1b2c3d4e5f6789abcdef"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Title, description, and validTo are required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      },
      "get": {
        "tags": [
          "Jobs - Admin"
        ],
        "summary": "Get all jobs (Admin view)",
        "description": "Retrieve all jobs regardless of status (ACTIVE, INACTIVE).\nIncludes jobs that are expired or not yet valid.\n\n**Supports filtering by job status:**\n- Filter by ACTIVE jobs only\n- Filter by INACTIVE jobs only\n- No filter returns all jobs\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Jobs retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "description": "Total number of jobs matching filter",
                          "example": 5
                        },
                        "filters": {
                          "type": "object",
                          "properties": {
                            "status": {
                              "type": "string",
                              "nullable": true,
                              "example": "ACTIVE"
                            }
                          },
                          "description": "Applied filters"
                        },
                        "jobs": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "example": "677a1b2c3d4e5f6789abcdef"
                              },
                              "title": {
                                "type": "string",
                                "example": "Frontend Developer"
                              },
                              "description": {
                                "type": "string",
                                "example": "We are hiring a React developer..."
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "ACTIVE",
                                  "INACTIVE"
                                ],
                                "example": "ACTIVE"
                              },
                              "validFrom": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-12T10:00:00Z"
                              },
                              "validTo": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-31T23:59:59Z"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-04T14:30:00Z"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "allJobs": {
                    "summary": "All jobs (no filter)",
                    "value": {
                      "success": true,
                      "message": "Jobs retrieved successfully",
                      "data": {
                        "total": 10,
                        "filters": {
                          "status": null
                        },
                        "jobs": []
                      }
                    }
                  },
                  "activeOnly": {
                    "summary": "Active jobs only",
                    "value": {
                      "success": true,
                      "message": "Jobs retrieved successfully",
                      "data": {
                        "total": 5,
                        "filters": {
                          "status": "ACTIVE"
                        },
                        "jobs": []
                      }
                    }
                  },
                  "inactiveOnly": {
                    "summary": "Inactive jobs only",
                    "value": {
                      "success": true,
                      "message": "Jobs retrieved successfully",
                      "data": {
                        "total": 3,
                        "filters": {
                          "status": "INACTIVE"
                        },
                        "jobs": []
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            },
            "description": "Filter jobs by status",
            "examples": {
              "active": {
                "summary": "Get only active jobs",
                "value": "ACTIVE"
              },
              "inactive": {
                "summary": "Get only inactive jobs",
                "value": "INACTIVE"
              }
            }
          }
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Admin login",
        "description": "Authenticate admin user and receive JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Invalid email or password"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}